+++ src/argilla/client/feedback/dataset/helpers.py
@@ -16,7 +16,6 @@ import typing
 from typing import Any, Dict, List, Optional, Type, Union
 
 import httpx
-from pydantic import BaseModel, Extra, ValidationError, create_model
 
 from argilla.client.feedback.constants import FIELD_TYPE_TO_PYTHON_TYPE
 from argilla.client.feedback.dataset.base import FeedbackDatasetBase
@@ -27,6 +26,7 @@ from argilla.client.sdk.v1.datasets import api as datasets_api_v1
 from argilla.client.sdk.v1.datasets.models import FeedbackDatasetModel
 from argilla.client.singleton import active_client
 from argilla.client.workspaces import Workspace
+from argilla.pydantic_v1 import BaseModel, Extra, ValidationError, create_model
 
 if typing.TYPE_CHECKING:
     from argilla.client.feedback.schemas.types import (