+++ src/argilla/client/login.py
@@ -15,10 +15,9 @@
 from pathlib import Path
 from typing import Dict, Optional
 
-from pydantic import AnyHttpUrl, BaseModel
-
 from argilla.client.sdk.commons.errors import HttpResponseError, UnauthorizedApiError
 from argilla.client.singleton import init
+from argilla.pydantic_v1 import AnyHttpUrl, BaseModel
 
 ARGILLA_CACHE_DIR = Path.home() / ".cache" / "argilla"
 ARGILLA_CREDENTIALS_FILE = ARGILLA_CACHE_DIR / "credentials.json"