+++ src/argilla/server/apis/v0/handlers/records.py
@@ -15,7 +15,6 @@
 from typing import Optional, Union
 
 from fastapi import APIRouter, Depends, Query, Security
-from pydantic import BaseModel
 
 from argilla.client.sdk.token_classification.models import TokenClassificationQuery
 from argilla.server.apis.v0.models.commons.params import CommonTaskHandlerDependencies
@@ -24,6 +23,7 @@ from argilla.server.apis.v0.models.text_classification import TextClassification
 from argilla.server.apis.v0.models.token_classification import TokenClassificationRecord
 from argilla.server.commons.config import TasksFactory
 from argilla.server.models import User
+from argilla.server.pydantic_v1 import BaseModel
 from argilla.server.security import auth
 from argilla.server.services.datasets import DatasetsService
 from argilla.server.services.search.service import SearchRecordsService