+++ src/argilla/server/daos/models/records.py
@@ -18,14 +18,13 @@ import warnings
 from datetime import datetime
 from typing import Any, Dict, Generic, List, Optional, TypeVar, Union
 
-from pydantic import BaseModel, Field, conint, constr, root_validator, validator
-from pydantic.generics import GenericModel
-
 from argilla import _messages
 from argilla._constants import _JS_MAX_SAFE_INTEGER, PROTECTED_METADATA_FIELD_PREFIX
 from argilla.server.commons.models import PredictionStatus, TaskStatus, TaskType
 from argilla.server.daos.backend.search.model import BaseRecordsQuery, SortConfig
 from argilla.server.helpers import flatten_dict
+from argilla.server.pydantic_v1 import BaseModel, Field, conint, constr, root_validator, validator
+from argilla.server.pydantic_v1.generics import GenericModel
 from argilla.server.settings import settings
 from argilla.utils import limit_value_length
 
@@ -108,7 +107,7 @@ class BaseRecordInDB(GenericModel, Generic[AnnotationDB]):
 
         return values
 
-    @root_validator()
+    @root_validator(skip_on_failure=True)
     def prepare_record_for_db(cls, values):
         values = cls.update_annotation(values, "prediction")
         values = cls.update_annotation(values, "annotation")