+++ tests/unit/client/feedback/schemas/test_questions.py
@@ -24,7 +24,8 @@ from argilla.client.feedback.schemas.questions import (
     TextQuestion,
     _LabelQuestion,
 )
-from pydantic import ValidationError
+
+from tests.pydantic_v1 import ValidationError
 
 
 @pytest.mark.parametrize(